Fix #12416: Refactor CURRENCY_AFFIX to use strong enum

+++ src/openrct2/config/IniWriter.hpp
@@ -34,7 +34,7 @@ interface IIniWriter
         std::string key = configEnum.GetName(value);
         if (key.empty())
         {
-            WriteInt32(name, value);
+            WriteInt32(name, static_cast<int32_t>(value));
         }
         else
         {
